% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{attrs}
\alias{attrs}
\title{Creates a vector of strings from variable names}
\usage{
attrs(...)
}
\arguments{
\item{...}{Names of the variables.}
}
\description{
Convert a set of variable names in a vector of strings with their names.
This function is useful to name the attributes to be read from CSV files.
}
