% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{buildRelations}
\alias{buildRelations}
\title{Overlap between two polygonal simple features}
\usage{
buildRelations(geom, ref, id)
}
\arguments{
\item{geom}{First simple feature, whose unique identified is the attribute ID.
This object is usually from GLOBIOM, a SimU, CR, or LU dataset.}

\item{ref}{Second simple feature, representing the region(s) to be mapped to
GLOBIOM representation.}

\item{id}{The attribute to be used as unique identifier for the second
simple feature.}
}
\description{
Computes the overlap between each pair of objects from two
simple features and returns their intersection area. This function can be
used to compute the relation between GLOBIOM representations and specific
regions such as states or biomes. It returns a data.table with three columns:
One ID from the first simple feature, one ID from the second simple feature,
and their intersection area in Mha. Objects with no intersection will not be
shown in the output.
}
