% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{equalSteps}
\alias{equalSteps}
\title{Creates intervals with equal sizes to be used by plot functions}
\usage{
equalslices(0, 100, 5)
}
\arguments{
\item{min}{The minimum value}

\item{max}{The maximum value}

\item{slices}{The number of slices to be created}

\item{data}{Any data that can be converted to a data.frame. The first column
of the data will be used to compute min and max.}
}
\description{
Creates intervals with equal sizes from min/max values to be
used by plot functions. If data is provided, it gets min/max from data.
}
