% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{filterGreater}
\alias{filterGreater}
\title{Filter the relations with greater area}
\usage{
filterGreater(data)
}
\arguments{
\item{data}{A tibble comming from normalizeRelationsByArea().}
}
\description{
Filter the relations with the greater intersection area. The
output will have only one row per each SimU, CR, or LU.
}
