% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{getCountries}
\alias{getCountries}
\title{Return all country names available for g2006.}
\usage{
getCountries(dataDirectory)
}
\arguments{
\item{dataDirectory}{Directory where input data is located. This directory needs to have
files available at https://www.dropbox.com/sh/sqocqe45jwmug2p/AAAbv-IAg24a_R4vYsP9zqV_a?dl=0.}
}
\description{
Return a vector of strings with all countries.
}
