% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{getCR}
\alias{getCR}
\title{Return CR geometries for a given country.}
\usage{
getCR(countryNames, dataDirectory, cache = TRUE)
}
\arguments{
\item{countryNames}{Names of the countries as a string vector. It can have one or more values.}

\item{dataDirectory}{Directory where input data is located. This directory needs to have
files available at https://www.dropbox.com/sh/sqocqe45jwmug2p/AAAbv-IAg24a_R4vYsP9zqV_a?dl=0.}

\item{cache}{If true (default), use the LUs precomputed.
Otherwise, it will compute from the original data.}
}
\description{
Return all ColRows of a given country as a set of simple feature
polygons, using the geometries of their respective SimUs.
}
