% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{getLU}
\alias{getLU}
\title{Return LU geometries for a given country.}
\usage{
getLU(countryName, dataDirectory)
}
\arguments{
\item{countryName}{Name of the country.}

\item{dataDirectory}{Directory where input data is located. This directory needs to have
files available at https://www.dropbox.com/sh/sqocqe45jwmug2p/AAAbv-IAg24a_R4vYsP9zqV_a?dl=0.}
}
\description{
Return all Large Units of a given country as a simple feature.
}
