% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getScenarios}
\alias{getScenarios}
\title{Return the scenarios of a directory}
\usage{
getScenarios(directory)
}
\arguments{
\item{directory}{Name of the directory}
}
\description{
This function returns the scenarios of a given directory. Each
scenario is stored as a directory.
It only removes the directory "results" from the list of directories if it exists.
}
