% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{getSimU}
\alias{getSimU}
\title{Return SimU geometries for a given country.}
\usage{
getSimU(countryNames, dataDirectory, join = TRUE, simu = NULL)
}
\arguments{
\item{countryNames}{Names of the countries as a string vector. It can have one or more values.}

\item{dataDirectory}{Directory where input data is located. This directory needs to have
files available at https://bit.ly/2RrgZi9 (shortened from https://www.dropbox.com/sh/sqocqe45jwmug2p/AAAbv-IAg24a_R4vYsP9zqV_a?dl=0).}

\item{join}{Should all SimuS with the same ID be represented together as a single MultiPolygon?
The default value is true.}

\item{simu}{A simple feature with loaded SimUs. Default is NULL, which means that the SimU data will be loaded from dataDirectory.}
}
\description{
Return all Simulation Units of a given country as a simple feature.
}
