% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{normalizeRelationsByArea}
\alias{normalizeRelationsByArea}
\title{Normalize relations according to the area}
\usage{
normalizeRelationsByArea(data)
}
\arguments{
\item{data}{A tibble coming from buildRelations().}
}
\description{
Normalize the relations according to the covered area. This
function is useful to distribute SimU, CR, or LU to other polygonal
representations. The returned values equal to 1 are completely distributed
to a single polygon. Whenever a value is less than 1, the SimU, CR, or LU
overlaps with two or more polygons, meaning that the content will be
divided to the respective polygons according to the intersection area.
}
