% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotAll}
\alias{plotAll}
\title{Plot several maps according to a prefix}
\usage{
plotAll(data, prefix, title = prefix, slices = 5, palette = "RdPu",
  additional = NULL, extension = "pdf", ...)
}
\arguments{
\item{data}{A sf (simple feature) object.}

\item{prefix}{A string with prefix name. Every attribute starting with
this prefix will be drawn in a separate map.}

\item{title}{A title prefix to be used in the map. The map title
will also have the year.}

\item{slices}{The number of slices. The default value is 5.}

\item{palette}{A string with an RColorBrewer palette.}

\item{additional}{One or more tmap functions (concatenated by +) to be
drawn after plotting the map.}

\item{extension}{The file extension. The default is "pdf".}

\item{...}{Any additional parameter to be passed to tm_fill.}
}
\description{
Plot several maps according to an attribute prefix.
This function is useful when the user wants to plot different
attributes using the same legend.
Tipically, the strings after the prefixes are years and the rest of
the documentation uses "year" with this in mind, but the
implementation is generic.
}
