% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processFile}
\alias{processFile}
\title{Convert a given output csv file to a shapefile.}
\usage{
processFile(shapefile, csvfile, description, outputFile = NULL,
  convertList = NULL)
}
\arguments{
\item{csvfile}{l The output csv file to be exported.}

\item{description}{A vector of strings with the attribute names.}

\item{outputFile}{Name of the shapefile to be saved.}

\item{convertList}{A list whose indexes are attribute values from the csv file and whose
values are going to replace the attribute values from the csv file. Both attributes and
values must be string.}

\item{datafile}{Name of the input shapefile. This file must be created using getCR,
getLU, or getSimU.}
}
\description{
Process a fiven csv and convert it to a shapefile.
}
