% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{readCR}
\alias{readCR}
\title{Read CR from a CSV file to tibble}
\usage{
readCR(directory, product)
}
\arguments{
\item{directory}{Path where the data is stored. Note that the last directory name will be the scenario's name.}

\item{product}{Name of the product to be read. The file name will be 'product_scenario.CSV'}
}
\description{
Read a CSV file as a tibble. The column names are identified automatically by
colrow package.
}
