% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{renameShape}
\alias{renameShape}
\title{Rename shapefile}
\usage{
renameShape(filename, newname)
}
\arguments{
\item{filename}{The file to be renamed, with .shp}

\item{newname}{The new name of the file, without .shp}
}
\description{
Rename a shapefile and all other files with the same name
}
\examples{
\donttest{
renameShape("myshape.shp", "my_new_shape")
}
}
