% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeAllColRow}
\alias{writeAllColRow}
\title{Write ID objects into a text file that can be used by GLOBIOM}
\usage{
writeAllColRow(tokens, header, outfile)
}
\arguments{
\item{tokens}{A vector of IDs.}

\item{header}{A string.}

\item{outfile}{Name of the file to be saved.}
}
\description{
Write 10 ID objects per line, plus a header and a line with slash,
according to GAMS requisites.
}
