% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{writeRelationsGMS}
\alias{writeRelationsGMS}
\title{Write relations to gms data}
\usage{
writeRelationsGMS(data, country, representation, attr, filename)
}
\arguments{
\item{data}{A tibble coming from normalizeRelationsByArea() or filterGreater().}

\item{country}{The country name.}

\item{representation}{A string that can be "SimU", "CR", or "LU".}

\item{attr}{A string with the attribute name.}

\item{filename}{The output filename.}
}
\description{
Write the relations to a gms data to be read by GAMS.
}
