% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processDirectory}
\alias{processDirectory}
\title{Process a given directory with results from different scenarios}
\usage{
processDirectory(directory, output = paste0(directory, "/results"))
}
\description{
Each scenario is represented as a directory, and its results
are CSV files within them.
}
