% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processDirectory}
\alias{processDirectory}
\title{Process a given directory with results from different scenarios}
\usage{
processDirectory(shapefile, directory, output = directory)
}
\arguments{
\item{shapefile}{A string with the shapefile that has all the CRs of
the output data. For each scenario, one shapefile will be created
containing this file plus the output of the respective scenario.}

\item{directory}{The directory where the scenarios are stored.}

\item{output}{The directory where the output is going to be saved. It
automatically appends "/results" to this directory. As default, the
output will be in the same directory where the scenarios are stored.}
}
\description{
Each scenario is represented as a directory, and its results
are CSV files within them.
}
