% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processFile}
\alias{processFile}
\title{Convert a given output csv file to a shapefile.}
\usage{
processFile(datafile, csvfile, outputFile, description, convertList = NULL)
}
\arguments{
\item{datafile}{Name of the input shapefile. This file must be created using getCR,
getLU, or getSimU.}

\item{csvfile}{l The output csv file to be exported.}

\item{outputFile}{Name of the shapefile to be saved.}

\item{description}{A vector of strings with the attribute names.}

\item{convertList}{A list whose indexes are attribute values from the csv file and whose
values are going to replace the attribute values from the csv file. Both attributes and
values must be string.}
}
