% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processFile}
\alias{processFile}
\title{Convert a given output csv file to a shapefile.}
\usage{
processFile(
  shapefile,
  csvfile,
  description,
  outputfile = NULL,
  convertList = NULL,
  aggregate = sum,
  toAggregate = NULL
)
}
\arguments{
\item{shapefile}{Name of the input shapefile. This file must be created using getCR,
getLU, or getSimU.}

\item{csvfile}{l The output csv file to be exported.}

\item{description}{A vector of strings with the attribute names.
It is possible to use attrs() to help creating this description.}

\item{outputfile}{Name of the shapefile to be saved.}

\item{convertList}{A list whose indexes are attribute values from the csv file and whose
values are going to replace the attribute values from the csv file. Both attributes and
values must be string.}

\item{aggregate}{A function to be used to join values with the same ID and attributes
to be joinded. As default, it will use sum, but it is possible to use functions such as
average.}

\item{toAggregate}{Attributes to be used only to aggregate, ignoring their values in
the created attribute names.}
}
\description{
Process a fiven csv and convert it to a shapefile.
}
